Backend/requirements.txt
# Core Flask Application
Flask==2.3.3
Flask-CORS==4.0.0

# Computer Vision and Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
scikit-image==0.21.0

# Deep Learning Frameworks
torch==2.0.1
torchvision==0.15.2
tensorflow==2.13.0

# Face Recognition
facenet-pytorch==2.5.3

# Object Detection
ultralytics==8.0.186

# OCR (Optical Character Recognition)
easyocr==1.7.0

# Scientific Computing
numpy==1.24.3
scipy==1.11.2

# Data Visualization (optional, for debugging)
matplotlib==3.7.2

# HTTP Requests (for model downloads)
requests==2.31.0

# Progress Bars
tqdm==4.66.1

# CUDA Support (optional, install based on your system)
# torch-audio==2.0.2  # Uncomment if needed
# torchtext==0.15.2   # Uncomment if needed

# Development Dependencies (optional)
# pytest==7.4.2
# pytest-cov==4.1.0
# flake8==6.0.0
# black==23.7.0

# Note: For GPU support, install PyTorch with CUDA support
# For Linux, you may need additional OpenCV system dependencies